+++ ui/main_window.py
@@ -11,6 +11,7 @@ from ui.pages.music_page import MusicPage
 from ui.pages.settings_page import SettingsPage
 from ui.pages.system_page import SystemPage
 from ui.pages.remote_page import RemotePage
+from ui.pages.modes_page import ModesPage
 from ui import theme_manager
 
 
@@ -48,12 +49,14 @@ class MainWindow(QMainWindow):
         self.pagina_config = SettingsPage()
         self.pagina_sistema = SystemPage()
         self.pagina_remoto = RemotePage()
+        self.pagina_modos = ModesPage()
 
         self.paginas.addWidget(self.pagina_inicio)
         self.paginas.addWidget(self.pagina_musica)
         self.paginas.addWidget(self.pagina_config)
         self.paginas.addWidget(self.pagina_sistema)
         self.paginas.addWidget(self.pagina_remoto)
+        self.paginas.addWidget(self.pagina_modos)
 
         # mapa chave -> página, pra achar rápido quando a sidebar avisar
         self._mapa_paginas = {
@@ -62,6 +65,7 @@ class MainWindow(QMainWindow):
             "config": self.pagina_config,
             "sistema": self.pagina_sistema,
             "remoto": self.pagina_remoto,
+            "modos": self.pagina_modos,
         }
 
         layout.addWidget(self.paginas)